Moved shim to separate bundle to remove automatic dependency to mongodb package
Now to import Mongo client you'll need to do it through '@event-driven-io/Pongo/shim'
