+++ src/packages/pongo/src/index.ts
@@ -1,3 +1,2 @@
 export * from './core';
-export * from './mongo';
 export * from './postgres';
